Git commit to add a lookahead function for a repeating waveform.
If a waveform is known to be repeating, the method checks if a newly found curve fits the next curve in the waveform before it looks for a new waveform pattern.

This reduces workload as a new waveform will not be found for every found curve (which will simply be the same waveform but with a different starting point along the oscillation).

Other changes that came about from this addition is the __is_valid property for curves and waveforms. This acts as a fail-safe for testing curves and waveforms that have not been correctly defined, which could lead to segfaults or zero division errors. Also, the curve functions have been condensed into a single method for brevity.

There are also minor changes to the naming of variables and commenting.